change functions and test to Credit files and add test if function fails
