+++ tests/mocks/server/handlers/credit.ts
@@ -1,6 +1,6 @@
 import { compose, rest as mockServer } from 'msw'
 import { PayWithFlexbase } from '../../../../src/index';
-import { badApiKey, badCompanyId, errorApiKey, errorCompanyId, goodApiKey, mockUrl } from '../constants';
+import { badApiKey, badCompanyId, errorApiKey, errorCompanyId, goodApiKey, goodCardId, goodCompanyId, goodUserId, mockUrl } from '../constants';
 
 export const credit_handlers = [
     mockServer.get(mockUrl + "/servicing/minimumDue", (request, response, context) => {
@@ -36,6 +36,37 @@ export const credit_handlers = [
 
         return response(res);
     }),
+
+    mockServer.post(mockUrl + "/servicing/payments/stripe", (request, response, context) => {
+
+        if(request.body && Object.keys(request.body).length === 0) {
+            const res = compose(
+                context.status(400),
+            );
+            return response(res);
+        }
+
+
+        const res = compose(
+            context.status(200),
+            context.json({
+                success: true,
+                cardPayment: {
+                    amount: "-1.53",
+                    asOf: "2022-07-28 21:50:40.482+00",
+                    byUser: goodUserId,
+                    companyId: goodCompanyId,
+                    datePosted: "2022-07-28 21:50:40.482+00",
+                    failureReason: null,
+                    id: goodCardId,
+                    status: "succeeded",
+                },
+            }),
+
+        );
+        return response(res);
+    }),
+
 
     mockServer.post<PayWithFlexbase>(mockUrl + "/credit/buyNow", (request, response, context) => {
